src/controleGrupos/Grupo.java
package controleGrupos;

import static util.Verificacoes.*;

public class Grupo {

	
	
	private String nome;

	
	
	public Grupo(String nome) {
		verificaStringInvalida(nome);
		verificaNull(nome);
		this.nome = nome;
	}

	
	
	public String getNome() {
		return this.nome;
	}



	/* (non-Javadoc)
	 * @see java.lang.Object#hashCode()
	 */
	@Override
	public int hashCode() {
		final int prime = 31;
		int result = 1;
		result = prime * result + ((this.getNome() == null) ? 0 : this.getNome().hashCode());
		return result;
	}



	/* (non-Javadoc)
	 * @see java.lang.Object#equals(java.lang.Object)
	 */
	@Override
	public boolean equals(Object obj) {
		if (this == obj)
			return true;
		if (obj == null)
			return false;
		if (getClass() != obj.getClass())
			return false;
		Grupo other = (Grupo) obj;
		if (this.getNome() == null) {
			if (other.getNome() != null)
				return false;
		} else if (!this.getNome().equals(other.getNome()))
			return false;
		return true;
	}

	
	
	
}
